Bump version of kotlin to 1.4
Notable change #1:
We had to drop spring-dependency-management plugin along the
way and specify versions of all spring-related dependencies manually.

This is because it somehow messes with 'exclude'-directives, which are
critical in this build to make sure that the right version of
kotlin-stdlib is bundled into playground. With that plugin,
'exclude'-directive is essentially ignored, and older stdlibs are picked
due to transitive dependencies.

See similar issue and response from the Gradle Team here:
https://discuss.gradle.org/t/excluded-dependence-comes-back-when-spring-boot-plugin-is-applied/17945/2

--

Notable change #2
We also had to fix versions of jackson-modulse in exeuctors.policy.
Previously, spring-plugin was forcing them to be 2.10.0, now they are
properly chosen to be 2.9.10 as specified in build-script